# internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: WaCalls API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    HTTP API for WaCalls, a self-hosted WhatsApp calling bridge. All /api/*
    endpoints require the bearer token configured via WACALLS_API_TOKEN
    (when unset, authentication is disabled). The SSE stream also accepts the
    token as the access_token query parameter. This document is the canonical
    contract: a test gate fails the build when routes and spec drift apart.
servers:
  - url: /
security:
  - bearerAuth: []
  - queryToken: []
paths:
...
  /api/sessions/{sid}/calls:
    post:
      operationId: startCall
      summary: Start an outgoing call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: E.164-ish phone number; non-digits are stripped
      responses:
        "200": { $ref: "#/components/responses/CallRef" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Operator identified by X-Client-Id is already on a call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Per-session concurrent call cap reached, or rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: Session is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listCalls
      summary: Live calls of the session, same shape as the SSE call-list
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "200":
          description: Live calls ordered by startedAt then callId
          content:
            application/json:
              schema:
                type: object
                required: [calls]
                properties:
                  calls:
                    type: array
                    items:
                      $ref: "#/components/schemas/CallRecord"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}:
    get:
      operationId: getCall
      summary: One live call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: The live call; 404 once it ends or if it belongs to another session
          content:
            application/json:
              schema:
                type: object
                required: [call]
                properties:
                  call:
                    $ref: "#/components/schemas/CallRecord"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
    delete:
      operationId: endCall
      summary: End an active call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "204":
          description: Call ended (idempotent)
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/sessions/{sid}/calls/{id}/accept:
    post:
      operationId: acceptCall
      summary: Accept an incoming call, claiming it for the X-Client-Id operator
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200": { $ref: "#/components/responses/CallRef" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Claimed by another client, or operator already on a call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/reject:
    post:
      operationId: rejectCall
      summary: Reject an incoming call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: Rejected (idempotent when the call is already gone)
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Call is in a state that cannot be rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}/video:
    post:
      operationId: enableVideo
      summary: Turn the local camera on or off mid-call
      description: >
        With no body (or any action other than "stop"), sends a WhatsApp mid-call
        video-upgrade request (<video state=3>) to the peer and derives the video
        SSRCs so the relay forwards both video streams; the browser starts
        capturing/encoding once this returns. With {"action":"stop"}, signals
        <video state=6> so the peer's UI drops our tile instead of freezing on the
        last decoded frame — only the outbound direction stops, inbound video and the
        relay subscription are left alone so a later enable needs no renegotiation.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [enable, stop]
                  default: enable
                  description: Anything other than "stop" requests the video upgrade
      responses:
        "200":
          description: Video state change signalled
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/hold:
    post:
      operationId: holdCall
      summary: Park the WhatsApp peer on hold music, or resume the call
      description: >
        On hold, the operator's microphone stops being relayed and hold music is
        injected into the WhatsApp leg instead. The music is read per hold from
        DataDir/hold-music.wav when present (uploaded via POST /api/holdmusic),
        otherwise the built-in chime is used. Broadcasts call-hold-state over SSE
        so every operator's UI shows the badge.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [hold, resume]
                  default: hold
                  description: Anything other than "resume" holds the call
      responses:
        "200":
          description: Hold state applied
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/transfer:
    post:
      operationId: transferCall
      summary: Blind-transfer a live call to another operator
      description: >
        Transfers between operators on the SAME paired WhatsApp account: the
        WhatsApp leg (CallManager) stays up, the call goes on hold so the peer
        hears music, and a call-transfer SSE event prompts the target operator.
        When that operator attaches its browser via POST .../webrtc, the new
        bridge replaces the old one and the call resumes automatically.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to:
                  type: string
                  description: Target operator identity (the X-Client-Id it connects with)
      responses:
        "200":
          description: Transfer offered to the target operator
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/recordings:
    get:
      operationId: listRecordings
      summary: List call recordings on disk
      description: >
        Lists the WAVs in RecordingsDir, newest first. Recordings are keyed by call
        ID across every session, since that is how they are stored. A recording whose
        call is still in progress is flagged active: its WAV header still carries the
        placeholder zero length, so it will not play until the call ends.
      responses:
        "200":
          description: Recordings on disk
          content:
            application/json:
              schema:
                type: object
                required: [recordings]
                properties:
                  recordings:
                    type: array
                    items:
                      type: object
                      required: [callId, size, modifiedAt, active]
                      properties:
                        callId:
                          type: string
                        size:
                          type: integer
                          description: Size in bytes
                        modifiedAt:
                          type: integer
                          format: int64
                          description: Last write time in milliseconds since the epoch
                        active:
                          type: boolean
                          description: Still being recorded, so not yet playable
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    queryToken:
      type: apiKey
      in: query
      name: access_token
  parameters:
    Sid:
      name: sid
      in: path
      required: true
      schema:
        type: string
      description: Session id
    CallId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Call id
    ClientId:
      name: X-Client-Id
      in: header
      required: false
      schema:
        type: string
      description: Operator identity; omitted means no ownership is recorded
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Per-IP rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such session or call
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request parameter or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    CallRef:
      description: Reference to the affected call
      content:
        application/json:
          schema:
            type: object
            required: [call]
            properties:
              call:
                type: object
                required: [callId]
                properties:
                  callId:
                    type: string
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CallStatus:
      type: string
      enum: [starting, ringing, connected, reconnecting, ended]
...
    CallRecord:
      type: object
      required: [sessionId, callId, owner, direction, peer, startedAt, status]
      properties:
        sessionId:
          type: string
        callId:
          type: string
        owner:
          type: [string, "null"]
          description: Claiming operator's client id; null when unclaimed, never ""
        direction:
          type: string
          enum: [outbound, inbound]
        peer:
          type: string
        peerName:
          type: string
          description: Resolved contact name for the peer, when known
        peerPhotoUrl:
          type: string
          description: Cached WhatsApp profile-picture URL for the peer, when known
        startedAt:
          type: integer
          format: int64
          description: Unix milliseconds
        status:
          $ref: "#/components/schemas/CallStatus"
        endedAt:
          type: integer
          format: int64
          description: Unix milliseconds; absent while the call is live
        endReason:
          type: string
          description: Absent while the call is live
...